+++ bp_home/__init__.py
@@ -1,0 +1,2 @@
+#Necessary to import files from that folder, marks it as "safe/available"
+#Also used in case one wants to init stuffs before loading anything from this folder... ¯\_ (ツ)_/¯
